update profile and drop course function

+++ api/controllers/enrollments.js
@@ -40,7 +40,7 @@ export const SelectCourse = (req, res) => {
         if (data.length) return res.status(409).json("Already enrolled in this course.");
 
         const q = "SELECT current_capacity, capacity FROM courses WHERE cid = ?";
-        
+
         db.query(q, [req.body.cid], (err, data) => {
           if (err) return res.status(500).json(err);
           if (data[0].current_capacity == data[0].capacity) return res.status(409).json("This course is already fulled.");
@@ -63,17 +63,23 @@ export const SelectCourse = (req, res) => {
 };
 
 export const DropEnrollment = (req, res) => {
+  const q = "SELECT * FROM enrollment_period";
+
+  db.query(q, (err, data) => {
+    if (err) return res.status(500).json(err);
+    if (data[0].canSelect == 0) return res.status(409).json("The enrollment period is closed.");
+
     const q = "DELETE FROM enrollment WHERE uid = ? AND cid = ?";
 
     db.query(q, [req.body.uid, req.body.cid], (err, data) => {
       if (err) return res.status(500).json(err);
-    });
-
-    const q2 = "UPDATE courses SET current_capacity = current_capacity-1 WHERE cid = ?";
-    db.query(q2, [req.body.cid], (err, data) => {
+      const q2 = "UPDATE courses SET current_capacity = current_capacity-1 WHERE cid = ?";
+      db.query(q2, [req.body.cid], (err, data) => {
         if (err) return res.status(500).json(err);
         return res.status(200).json("Drop successfully.");
-    })
+      })
+    });
+  });
 };
 
 export const GradeUpload = (req, res) => {
